refactor(reservation-domain): 좌석 점유 도메인 구조 정리
- SeatTimeSlotOccupancy를 SeatOccupancy로 이름 변경함.
- Reservation에서 좌석 점유 관계를 제거해 예약 상태와 점유 정보를 분리함.
